Fix connection growing issue in SMB2 path
The architecture of the SMB2 path is to define indefinite number of
SMB2Clients in a concurrent situation but when handing over just manage
a single reference by using a atomic variable and closing others.
Here the issue was we were closing the SMBClient yet since underlying
resources are utilized the connections were not closed properly.
Hence fixed that.

Fixes: https://github.com/wso2/micro-integrator/issues/3295